Refactor ServerMessage so it can be shared by ViewCrawler without too much circular dependency
