+++ src/androidTest/java/com/mixpanel/android/mpmetrics/DecideCheckerTest.java
@@ -4,6 +4,7 @@ import android.content.Context;
 import android.os.Bundle;
 import android.test.AndroidTestCase;
 
+import com.mixpanel.android.util.ServerMessage;
 import com.mixpanel.android.viewcrawler.UpdatesFromMixpanel;
 
 import org.json.JSONArray;
@@ -29,7 +30,7 @@ public class DecideCheckerTest extends AndroidTestCase {
         mDecideMessages3.setDistinctId("DISTINCT ID 3");
     }
 
-    public void testReadEmptyLists() throws ServiceUnavailableException {
+    public void testReadEmptyLists() throws ServerMessage.ServiceUnavailableException {
         mDecideChecker.addDecideCheck(mDecideMessages1);
 
         mPoster.response = bytes("{}");
@@ -50,7 +51,7 @@ public class DecideCheckerTest extends AndroidTestCase {
         });
     }
 
-    public void testReadSurvey1() throws ServiceUnavailableException {
+    public void testReadSurvey1() throws ServerMessage.ServiceUnavailableException {
         mDecideChecker.addDecideCheck(mDecideMessages1);
 
         mPoster.response = bytes(
@@ -94,7 +95,7 @@ public class DecideCheckerTest extends AndroidTestCase {
         assertEquals(textChoices.size(), 0);
     }
 
-    public void testReadSurvey2() throws ServiceUnavailableException {
+    public void testReadSurvey2() throws ServerMessage.ServiceUnavailableException {
         mDecideChecker.addDecideCheck(mDecideMessages1);
         mPoster.response = bytes(
                 "{\"surveys\":[{\"collections\":[{\"id\":151,\"selector\":\"\\\"@mixpanel\\\" in properties[\\\"$email\\\"]\"}],\"id\":299,\"questions\":[{\"prompt\":\"PROMPT1\",\"extra_data\":{\"$choices\":[\"Answer1,1\",\"Answer1,2\",\"Answer1,3\"]},\"type\":\"multiple_choice\",\"id\":287},{\"prompt\":\"How has the demo affected you?\",\"extra_data\":{\"$choices\":[\"I laughed, I cried, it was better than \\\"Cats\\\"\",\"I want to see it again, and again, and again.\"]},\"type\":\"multiple_choice\",\"id\":289}]}]}"
@@ -125,7 +126,7 @@ public class DecideCheckerTest extends AndroidTestCase {
         assertEquals(mcChoices.get(2), "Answer1,3");
     }
 
-    public void testBadDecideResponses() throws ServiceUnavailableException {
+    public void testBadDecideResponses() throws ServerMessage.ServiceUnavailableException {
         mDecideChecker.addDecideCheck(mDecideMessages1);
 
         // Corrupted or crazy responses.
@@ -191,7 +192,7 @@ public class DecideCheckerTest extends AndroidTestCase {
         mEventBinder.bindingsSeen.clear();
     }
 
-    public void testDecideHonorsFallbackDisabled() throws ServiceUnavailableException {
+    public void testDecideHonorsFallbackDisabled() throws ServerMessage.ServiceUnavailableException {
         mConfig.fallbackDisabled = false;
         mPoster.response = bytes("{\"surveys\":[], \"notifications\":[]}");
         mDecideChecker.addDecideCheck(mDecideMessages1);
